backend/pyproject.toml
[project]
name = "krait-backend"
version = "0.1.0"
description = "KRAIT - Advanced GPU kernel analysis and optimization platform with AI-powered code review"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.10.1",
    "python-dotenv>=1.0.0",
    "langchain>=0.1.0",
    "langchain-google-genai>=1.0.0",
    "langchain-openai>=0.0.5",
    "transformers>=4.36.0",
    "torch>=2.1.0",
    "accelerate>=0.25.0",
    "sentence-transformers>=2.2.2",
    "httpx>=0.25.0",
    "google-generativeai>=0.3.0",
    "openai>=1.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
